requirements.txt
flask
flask_sqlachemy
flask_marshmallow
requests
Pillow
numpy